+++ Tests/Rendering/SpriteAnimatorTests.cs
@@ -15,16 +15,16 @@
         [Category("Unit Tests")]
         public static void SpriteAnimator_LoopingTest() {
             var numberOfSteps = 3;
-            var animation = CreateAnimation(numberOfSteps, true);
-            var animator = CreateAnimator(animation, 1);
+            var animation = CreateAnimation(numberOfSteps);
+            var animator = CreateAnimator(1);
             animator.Initialize(Substitute.For<IScene>());
 
             var gameTime = new GameTime {
                 ElapsedGameTime = TimeSpan.FromMilliseconds(100d)
             };
 
-            Assert.AreEqual(animation.Steps.ElementAt(0).Sprite.Id, animator.Sprite.Id);
-
+            animator.Play(animation, true);
+            Assert.IsNull(animator.Sprite);
             animator.UpdateAsync(gameTime).Wait();
             Assert.AreEqual(animation.Steps.ElementAt(0).Sprite.Id, animator.Sprite.Id);
 
@@ -41,7 +41,7 @@
             Assert.AreEqual(animation.Steps.ElementAt(0).Sprite.Id, animator.Sprite.Id);
         }
 
-        private static SpriteAnimation CreateAnimation(int numberOfSteps, bool shouldLoop) {
+        private static SpriteAnimation CreateAnimation(int numberOfSteps) {
             var steps = new List<SpriteAnimationStep>();
             for (var i = 0; i < numberOfSteps; i++) {
                 steps.Add(new SpriteAnimationStep {
@@ -50,11 +50,11 @@
                 });
             }
 
-            return new SpriteAnimation(steps, shouldLoop);
+            return new SpriteAnimation(steps);
         }
 
-        private static SpriteAnimationComponent CreateAnimator(SpriteAnimation animation, int frameRate) {
-            var animator = new SpriteAnimationComponent(animation) {
+        private static SpriteAnimationComponent CreateAnimator(byte frameRate) {
+            var animator = new SpriteAnimationComponent() {
                 FrameRate = frameRate
             };